backend/app/config.py
# backend/app/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

BASE_DIR = Path(__file__).resolve().parents[1]
ENV_FILE = BASE_DIR / ".env"

# Local dev convenience: load .env if present (no crash if missing)
if ENV_FILE.exists():
    load_dotenv(dotenv_path=ENV_FILE)

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "").strip()
PUBLIC_BASE_URL = os.getenv("PUBLIC_BASE_URL", "").strip().rstrip("/")

# Do NOT raise here (Render pe env vars set honge)
# If missing, endpoints will show warnings / limited functionality.