backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
twilio==9.2.3
requests==2.32.3
websockets==12.0
openai==1.40.6
pydantic==2.8.2